% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.Functions.r
\name{Plot.AbundTime}
\alias{Plot.AbundTime}
\title{Plot Species Abundances Through Time}
\usage{
Plot.AbundTime(SimResult)
}
\arguments{
\item{SimResult}{A named list of data frames, one per recorded simulation timestep, each containing individuals alive at that timestep with a Species column. List names must be coercible to numeric simulation timestamps. Output of Sim.Compute().}
}
\value{
A ggplot object visualising the abundance of each species across simulation time as line plots.
}
\description{
This function visualises the abundance trajectory of each species across simulation time.
}
\examples{
data("SimulationOutput")

Plot.AbundTime(SimulationOutput)

}
\author{
Erik Kusch
}
