% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.Functions.r
\name{Plot.Environment}
\alias{Plot.Environment}
\title{Plot Environmental Matrix}
\usage{
Plot.Environment(Env_mat)
}
\arguments{
\item{Env_mat}{Environmental matrix as produced by Sim.Space(). Cells contain environmental values whereas column and row names contain spatial coordinates.}
}
\value{
A ggplot object visualising the (subsampled) environmental matrix as a heatmap.
}
\description{
This function creates a heatmap visualisation of an environmental matrix as produced by Sim.Space(). To keep rendering fast, only every ninth row and column of the matrix is plotted.
}
\examples{
data("Env_mat")

Plot.Environment(Env_mat)

}
\author{
Erik Kusch
}
