% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.Functions.r
\name{Plot.NetMat}
\alias{Plot.NetMat}
\title{Plot Association/Interaction Network Matrix}
\usage{
Plot.NetMat(Network_mat)
}
\arguments{
\item{Network_mat}{A matrix object with association/interaction strength stored as cell values. Output of Sim.Network().}
}
\value{
A ggplot object visualising the network matrix as a heatmap of pairwise association/interaction strengths.
}
\description{
This function creates a heatmap visualisation of an association/interaction network matrix as produced by Sim.Network().
}
\examples{
data("Network_mat")

Plot.NetMat(Network_mat)

}
\author{
Erik Kusch
}
