% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.Network.R
\name{Sim.Network}
\alias{Sim.Network}
\title{Generate Random Association/Interaction Network Matrices}
\usage{
Sim.Network(
  n_spec = 5,
  NetworkType = "Association",
  Sparcity = 0.5,
  MaxStrength = 1,
  seed = 42
)
}
\arguments{
\item{n_spec}{Numeric. Number of species / size of node set.}

\item{NetworkType}{Character. Whether to create an interaction or association matrix. In an interaction matrix, columns act on rows. In an association matrix, entries are mirrored around the diagonal. Possible values: "Association" or "Interaction".}

\item{Sparcity}{Numeric. Proportion of unique associations or interactions which are exactly 0.}

\item{MaxStrength}{Numeric. Maximum value of associations or interactions.}

\item{seed}{Numeric. Seed for random processes.}
}
\value{
An matrix object with association/interaction strength stored as in cells.
}
\description{
This function generates a random network matrix representing either an undirected (association) or directed (interaction) network matrix according to user specifications.
}
\examples{
net <- Sim.Network(
  n_spec = 20,
  NetworkType = "Association",
  Sparcity = 0.5,
  MaxStrength = 1,
  seed = 42
)
Plot.NetMat(net)

net <- Sim.Network(
  n_spec = 10,
  NetworkType = "Interaction",
  Sparcity = 0,
  MaxStrength = 20,
  seed = 21
)
Plot.NetMat(net)

}
\author{
Erik Kusch
}
