% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.Compute.R
\name{Sim.Compute}
\alias{Sim.Compute}
\title{Carry out Demographic Simulation}
\usage{
Sim.Compute(
  d0 = 0.4,
  b0 = 0.6,
  k_vec,
  ID_df,
  env.xy,
  env.sd = 2.5,
  mig.sd = 0.2,
  mig.top = 0.05,
  mig.trunc = 1,
  interac.maxdis = 0.5,
  interac.mat,
  interac.scale = 1,
  perturbs = NULL,
  Sim.t.max = 10,
  Sim.t.inter = 0.1,
  seed = 42,
  verbose = TRUE,
  RunName = "",
  writeFile = TRUE
)
}
\arguments{
\item{d0}{Numeric. background death rate (gets altered by the environment and interactions).}

\item{b0}{Numeric. background birth rate (remains constant).}

\item{k_vec}{Named vector containing carrying capacity for each species. Output of Sim.CarryingK().}

\item{ID_df}{Data frame of initialising individuals with columns ID, Trait, X, Y, and Species. Output of Sim.Initialise()$ID_df.}

\item{env.xy}{Environmental matrix as produced by Sim.Space(). Cells contain environmental values whereas column and row names contain spatial coordinates.}

\item{env.sd}{Numeric. Habitat suitability in death rate function. Higher values allow individuals to persist in areas of greater environmental maladaptation.}

\item{mig.sd}{Numeric. Standard deviation of 0-centred normal distribution from which natal dispersal is drawn.}

\item{mig.top}{Numeric. Distance around a birth event within which the chance of a birth to occur are high (i.e. the flat-top portion of a bivariate flat-top normal distribution).}

\item{mig.trunc}{Numeric. Maximum distance of migration.}

\item{interac.maxdis}{Numeric. Distance within which neighbouring individuals interact with a focal individual.}

\item{interac.mat}{An matrix object with association/interaction strength stored as cell values. Output of Sim.Network().}

\item{interac.scale}{Numeric. Scaling factor for strength of interaction effects on death rate.}

\item{perturbs}{Named vector of numeric values. Magnitude of perturbations applied to the system during the simulation. Named values correspond to the simulation time from which point on perturbation is applied.}

\item{Sim.t.max}{Numeric. Maximum simulation time.}

\item{Sim.t.inter}{Numeric. Interval length at which to record simulation outputs (measured in simulation time).}

\item{seed}{Numeric. Seed for random processes.}

\item{verbose}{Logical. Whether to print simulation time and sampling interval.}

\item{RunName}{Character. Name for temporary .RData object written to disk.}

\item{writeFile}{Logical. Whether temporary .RData should be written to disk.}
}
\value{
A list containing data frame object with the same columns as ID_df at each sampling interval defined via n_inter until Sim.t.max is reached.
}
\description{
This function carries out spatially explicit, individual-based demographic simulation for the generation of data products ready for network inference.
}
\examples{
data("Initialise_df")
data("CarryingK_vec")
data("Network_mat")
data("Env_mat")

SimResult <- Sim.Compute(
  # Demographic parameters
  d0 = 0.4,
  b0 = 0.6,
  k_vec = CarryingK_vec,
  ID_df = Initialise_df,

  # Spatial parameters
  env.xy = Env_mat,
  env.sd = 2.5,
  mig.sd = 0.2,
  mig.top = 0.05,

  # Interaction parameters
  interac.maxdis = 0.5,
  interac.mat = Network_mat,
  interac.scale = 1,

  # Simulation parameters
  Sim.t.max = 5,
  Sim.t.inter = 0.1,
  seed = 42,
  verbose = TRUE, # whether to print progress in time as current time
  RunName = "Trial"
)

SimResult <- Sim.Compute(
  # Demographic parameters
  d0 = 0.4,
  b0 = 0.6,
  k_vec = CarryingK_vec[1:5],
  ID_df = Initialise_df[Initialise_df$Sp \%in\% paste0("Sp_0", 1:5), ],
  # Spatial parameters
  env.xy = Env_mat,
  env.sd = 1.5,
  mig.sd = 0.25,
  mig.top = 0.15,
  # Perturbation parameters
  perturbs = c("100" = 2, "150" = 2),
  # Interaction parameters
  interac.maxdis = 0,
  interac.mat = Network_mat[1:5, 1:5],
  interac.scale = 0,
  # Simulation parameters
  Sim.t.max = 200,
  Sim.t.inter = 0.5,
  seed = 42,
  verbose = TRUE,
  RunName = "Perturbations",
  writeFile = FALSE
)
print(Plot.AbundTime(SimResult))
}
\author{
Erik Kusch, Natural History Museum, University of Oslo, Norway.
}
